finish is close

+++ ft_printf/ft_power.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-long long		ft_power(long long num, int power)
+unsigned long long	ft_power(long long num, int power)
 {
-	int	res;
+	unsigned long long	res;
 
 	res = num;
 	if (power <= 0)
